% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Dnorm}
\alias{Dnorm}
\title{Chi-Square Difference Test}
\usage{
Dnorm(msmall, mbig)
}
\arguments{
\item{msmall}{A \code{lavaan} object}

\item{mbig}{A \code{lavaan} object containing a model run nested within \code{msmall}. \code{mbig} is the model run with the larger chi-square.}
}
\value{
A data.frame containing the chi-square difference, degrees of freedom associated with the difference test, and a pvalue.
}
\description{
This function calculates the chi-square difference of two nested CFA models,
and tests its significance.
}
\examples{
ana.model<-'
f1=~ y1 + y2 + y3
f2=~ y4 + y5 + y6
'
m0<-cfa(ana.model,data=data1,group="group")
m1<-cfa(ana.model,data=data1,group="group",group.equal=c("loadings"))
Dnorm(m0,m1)
}

