% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DSB}
\alias{DSB}
\title{Scaled Chi-Square Difference Test}
\usage{
DSB(msmall, mbig)
}
\arguments{
\item{msmall}{A \code{lavaan} object}

\item{mbig}{A \code{lavaan} object containing a model run nested within \code{msmall}. \code{mbig} is the model run with the larger chi-square.}
}
\value{
A data.frame containing the scaled and unscaled chi-square difference, degrees of freedom associated with the difference test, and the pvalues associated with the scaled and unscaled tests.
}
\description{
This function calculates the Satorra-Bentler scaled chi-square difference of two nested CFA models,
and tests its significance. This difference test is robust to non-normality of data.
}
\examples{
ana.model<-'
f1=~ y1 + y2 + y3
f2=~ y4 + y5 + y6
'
m0<-cfa(ana.model,data=data1,group="group")
m1<-cfa(ana.model,data=data1,group="group",group.equal=c("loadings"))
DSB(m0,m1)
}

