% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{invariance}
\alias{invariance}
\alias{invariance-package}
\title{Measurement Invariance Testing}
\description{
Invariance centered difference testing using lavaan Structural Equation Models.
}
\details{
Functions are included for difference testing nested structural equation models.
Additionally, a function that automates the process of difference testing invariance models is included.
}

