% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invarianceSeries}
\alias{invarianceSeries}
\title{Run The Standard Invariance Series}
\usage{
invarianceSeries(model, data, group = "group")
}
\arguments{
\item{model}{A CFA model specified in \code{lavaan} syntax}

\item{data}{A data.frame}

\item{group}{A string identifying the grouping variable in \code{data}#'}
}
\value{
A data.frame containing the scaled and unscaled chi-square difference, degrees of freedom associated with the difference test, and the pvalues associated with the scaled and unscaled tests.
}
\description{
This function runs the series of difference tests typically run when evaluating measurement invariance for a psychometric instrument.
}
\examples{
ana.model<-'
f1=~ y1 + y2 + y3
f2=~ y4 + y5 + y6
'
invarianceSeries(ana.model,data)
}

