% Generated by roxygen2 (4.0.2): do not edit by hand
\name{m0run}
\alias{m0run}
\title{Configural Invariance Fit Test}
\usage{
m0run(model, data, group = "group")
}
\arguments{
\item{model}{A CFA model specified in \code{lavaan} syntax}

\item{data}{A data.frame}

\item{group}{A string identifying the grouping variable in \code{data}}
}
\value{
An object of class \code{lavaan}, for which several methods are available,
including a \code{summary} method.
}
\description{
This function calls the lavaan cfa function, and runs the test of
configural invariance on a multiple group dataset.
}
\details{
This function runs the first stage of the standard invariance series.
}
\examples{
ana.model<-'
f1=~ y1 + y2 + y3
f2=~ y4 + y5 + y6
'
m0<-m0run(ana.model,data)
summary(m0)
}

